requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
Flask==3.1.3
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pip==23.1.2
psycopg2==2.9.11
python-dotenv==1.2.2
requests==2.32.5
setuptools==65.5.0
SQLAlchemy==2.0.48
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
